+++ src/sql/cmd/snippets.py
@@ -26,7 +26,7 @@ def _modify_display_msg(key, remaining_keys, dependent_keys=None):
     if dependent_keys:
         msg = f"{msg}{', '.join(dependent_keys)} depend on {key}\n"
     if remaining_keys:
-        msg = f"{msg}Stored snippets : {', '.join(remaining_keys)}"
+        msg = f"{msg}Stored snippets: {', '.join(remaining_keys)}"
     else:
         msg = f"{msg}There are no stored snippets"
     return msg
@@ -104,6 +104,8 @@ def snippets(others):
         key = args.delete_force
         deps = store.get_key_dependents(key)
         remaining_keys = store.del_saved_key(key)
+        for dep in deps:
+            store.store[dep].remove_snippet_dependency(key)
         return _modify_display_msg(key, remaining_keys, deps)
 
     elif args.delete_force_all: