+++ src/tests/test_magic_cmd.py
@@ -789,6 +789,13 @@ def test_delete_invalid_snippet(arg, ip_snippets):
     assert str(excinfo.value) == "No such saved snippet found : non_existent_snippet"
 
 
+@pytest.mark.parametrize("arg", ["--delete-force", "-D"])
+def test_delete_snippet_when_dependency_force_deleted(ip_snippets, arg):
+    ip_snippets.run_cell(f"%sqlcmd snippets {arg} high_price")
+    out = ip_snippets.run_cell("%sqlcmd snippets --delete high_price_a").result
+    assert "high_price_a has been deleted.\nStored snippets: high_price_b" in out
+
+
 @pytest.mark.parametrize(
     "arguments", ["--table schema1.table1", "--table table1 --schema schema1"]
 )