+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.10.3dev
 
+* [Fix] Remove force deleted snippets from dependent snippet's `with` (#717)
+
 ## 0.10.2 (2023-09-22)
 
 * [Feature] Improved messages when loading configurations from `pyproject.toml` file.